shared/SBIR/code/requirements.txt
google-genai>=1.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0

